Initial import of sandboxfs
sandboxfs is an open-source FUSE file system that exposes a combination
of multiple files and directories from the host's file systems in the
form of a virtual tree with an arbitrary layout.

sandboxfs is designed to allow running sandboxed commands with limited
access to the file system by using the virtual tree as their new root,
and to do so consistently across a variety of platforms.

Project import generated by Copybara.

PiperOrigin-RevId: 163828980