Rewrite the CLI integration tests in Go.
This is the first step in rewriting all the integration tests in Go,
which is a requirement for open-sourcing the project (as the current
tests are written using gbash, and that is not open-source).

Note that there are a lot of new TODOs in the tests.  These are for
inconsistencies I've noticed in the code being checked while porting
the tests.  I intentionally don't want to modify the original code
until all the new tests are in place and in good shape.

Also note that this adds many more tests than there used to be for the
CLI, which is in part why I discovered the inconsistencies.

PiperOrigin-RevId: 164768709