Rewrite the reconfiguration tests in Go.
This CL attempts to translate the existing shell tests in Go, but as you will
notice from the TODOs, such conversion raises many questions on whether we
should or should not have the current static/dynamic duality.  I'd like to
address this, but it ought to be done separately once all tests are in place.

PiperOrigin-RevId: 165571604