Rewrite the nested mappings tests in Go.
This finalizes the rewrite of all the integration tests in Go and thus
deletes the remaining shell tests.

Note that, as part of this change, I'm adding a couple of tests that
don't really come from the old tests... but represent bugs that I
discovered while actually converting those tests.  Namely: file
truncations don't seem to work and attempting to mount sandboxfs on a
non-existing directory causes a crash instead of a controlled error.

As before, this just focuses on converting the tests, not on fixing
actual problems with the code.  Correcting the bugs will come later.

PiperOrigin-RevId: 165736474