Document sandboxfs and prepare for publication.
This change extends the documentation of sandboxfs by fully documenting
the command-line tool in its own manual page (traditional of low-level
Unix system utilities), and simplifies the contents of the README.md
to focus on the details that newcomers may be interested in.

Additionally, this adds a new NEWS.md file to collect the major release
notes between releases, and places a clear disclaimer that integration
points in the command-line are subject to change.

PiperOrigin-RevId: 166189219