Check error before defer commands that may fail.
If a file is opened, it returns the file pointer and error. However,
in case of an error, the file pointer would be nil. Thus, the error
needs to be checked, or else `defer file.Close()` may be executed on
a nil pointer causing the program to crash.
Same goes for other function calls like fuse.Mount().