Fix test for targets that are doubly-mapped on macOS.
FUSE's kernel-level cache for file contents seems to behave differently
between Linux and macOS: on Linux, the cache is flushed at file open time,
but on macOS that doesn't seem to be the case.  This inconsistency was
causing the testNesting_SameTarget test to fail on macOS but to pass on
Linux.

Fix this by making the behavior system-specific.  Even though we think
that the behavior on macOS is suboptimal, we know what the current
behavior is, so it's better to test for it in case the behavior changes.

PiperOrigin-RevId: 168601010