Do not rely on GOPATH to find the final sandboxfs binary.
GOPATH is intended to be a list of colon-separated directories and,
as such, a reference like ${GOPATH}/bin to find the location of the
installed binary is conceptually wrong.  The fact that it currently
works is only accidental because Travis is adding a single directory
to GOPATH.

To fix this, be explicit about where "go build" leaves the final
binary and then pass that to the integration tests.

Fix suggested by Pallav Agarwal.

PiperOrigin-RevId: 168698990