Simplify signal handling so that defers can run.
When signal handling was first introduced for controlled exits, we tried
to be smart by redelivering the caught signal to the process after
uninstalling the signal handler.  The goal was to cause the process to
terminate with the correct exit code from the operating system's point
of view.

Unfortunately, this approach had a side-effect: defers don't run when
signals were caught.  This was not a big problem because we were only
using defers for file handle cleanups... until now.  It turns out that,
to implement profiling, we need defers to work properly so this does
become a problem.

To fix this, simply report caught signals as errors.  The good news is
that this doesn't break Unix process termination semantics because...
well, they are broken anyway with Go, which insists on reporting signals
as "error code 1".

PiperOrigin-RevId: 170095677