Add profiling support.
This change adds two complementary features to sandboxfs: the ability to
record full-program CPU and memory profiles to disk (via the --cpu_profile
and --mem_profile flags), and the ability to expose HTTP pprof endpoints
(via the --listen_address flag).

Both are interesting for different use-cases.  The disk-based profiles
will be useful for Bazel integration, as Bazel will want to capture
profiles for the various (possibly short-lived) sandboxfs instances it
starts.  The HTTP profiles will be useful for interactive usage, which is
handy with long-lived sandboxfs instances started manually.

We will require these features to tune the performance of sandboxfs.

PiperOrigin-RevId: 170097876