Define flags only within the commands that need them.
Refactor the main program to not care about command-specific flags: it's
the responsibility of the command-handling functions to define the flags
they accept and to parse them.

As a side effect, this avoids the need to use flag.Lookup() throughout
the program as the flag values are directly available where needed.

No functional changes.

PiperOrigin-RevId: 170351385