Use the public flag and log packages.
sandboxfs intends to be primarily an open-source package that will later
be reimported into third_party/golang/sandboxfs/ once we have declared
GitHub to be authoritative.  As a result, the external copy will be the
main one and will have to use the public versions of the flag and log
packages.  (This should happen relatively soon, obsoleting the current
primary copy in third_party/sandboxfs/.)

For this reason, convert the code to use the external packages
unconditionally.  This has two immediate benefits: first, it simplifies
our export process, which ensures that the behavior of what we export
matches exactly what we currently test; and, second, it fixes a problem
with parsing flags in which sandboxfs was currently swallowing but
mostly ignoring the predefined google3-only flags.

I'm doing this now because while trying to address the TODOs in
cli_test.go regarding error reporting, I hit a dead end when trying to
unify how the external and the internal packages behave.  Making this
work consistently (if possible at all, which I'm not clear about yet)
would require some rather nasty hacks that I'd prefer to avoid by
simply switching our code to use the public versions upfront.

PiperOrigin-RevId: 170480544