Fix race between signal handler installation and mount point setup.
If a signal arrived between the installation of the corresponding signal
handler and the actual mount(2) system call, the signal would be lost
and the FUSE server loop would be entered, making sandboxfs remain in a
running (and possibly-broken) state.

To fix this, ensure that signals are installed before the mount point is
prepared, but only handle them after the mount point is ready.

PiperOrigin-RevId: 170511064