Handle the Setattr call to implement metadata updates.
This adds proper support for metadata changes such as file permissions,
ownerships, and times.

In particular, this also fixes a long-standing bug that caused short file
writes to not clear old longer content and also fixes the functionality
of the basic truncate system call.  (Why? Because the Setattr call is also
used to communicate file size changes, and we were not dealing with those
at all in the past.)

PiperOrigin-RevId: 170515859