Move integration testing utilities to their own package.
The miscellaneous utilities for integration testing in common.go have
grown quite a bit and it's becoming increasingly hard to reason about
them because of the huge "public" surface they have within tests.

To address this, move such utilities into a new "utils" package so that
we must clearly declare what's public and what's not.

As part of this change, I am purely doing symbol renames and not trying
to improve the abstractions offered by the common utilities.  Those will
come later.  Most renames are just to capitalize the symbol and make it
public, but in other cases I'm taking the chance to adopt the more
Go-accepted names (fooOrFatal to MustFoo).

PiperOrigin-RevId: 170858047