Implement support for allow_other and allow_root.
Add the ability to control who can access a mounted sandboxfs instance.
This is via a new --allow flag that can take three values: self, other,
and root.  self, the default, preserves the previous behavior of only
allowing the user that mounted the file system.  other and root
correspond to the allow_other and allow_root FUSE settings.

This is necessary because, on macOS, we need to mount sandboxes with
allow_other if we intend to run signed binaries (like Java) through
them.  Otherwise, the system is unable to validate their signatures and
the processes die immediately.  This took multiple hours to track down.

Also, this feature took about 15 minutes to implement... and multiple
hours to get the tests in place.  Mind you, writing the tests made me
discover ugly inconsistencies between Linux's FUSE and OSXFUSE, but
I'm not trying to address those.

PiperOrigin-RevId: 171512697