Fix handling of error message for --allow=root on Linux.
Last-minute changes on a previous CL caused the error handling and tests
for --allow=root on Linux to be bogus.  The reason is that I forgot to
properly run the modified tests as root (which currently requires some
hackery before exporting) and the tests failed as soon as we hit Travis.

First, the addition of the "known brokenness" note to this condition was
broken because of copy/paste mistakes when defining the new mountError
type.

Second, the tests were using the magic %ROOT% path on arguments to
sandboxfs, which is handled internally by utils.MountSetup*, but we are
not calling such functions for this specific test.

And third, there was a typo in the error message.

PiperOrigin-RevId: 171565194