Handle signals received while the mount point is busy.
If we receive a signal while there are open file handles on the mount point,
our cleanup process will fail because we cannot unmount it, and if this
fails, we enter a zombie-like state in which we try to exit but cannot and we
stop processing signals.  For additional fun, forcibly killing the process
at this point leaves a mount point behind on macOS (because of a kernel bug
most likely).

To fix this, make the signal handler repeat the unmount operation indefinitely
when it encounters an error so that the mount point goes away as soon as it
stops being busy.  This might look a bit strange but I think it's the best
we can do for now: ignoring the signal would expose a race to the caller
(whereby the caller would need to retry sending signals if it suspects file
descriptors are "still being cleaned up").

I'm not sure if there is any other solution though: we could try revoking all
open file descriptors on the file system somehow, but we don't have control
on those (not to mention that revoke(2) is a BSD-only operation).  Or we could
make the file system enter a form of "lame" mode in which new operations fail
and only closes continue to work... but that requires significant plumbing and
locking throughout that may be detrimental performance-wise.

PiperOrigin-RevId: 171813224