Fix a race between cache invalidations and directory lookups.
The cache invalidation code did not hold the per-directory lock while
reading the directory entries, which means that we could be iterating
on maps that are concurrently modified.  (Thankfully, Go catches that
and prints a descriptive message, which is how I originally noticed
this problem.)

Fix this by holding the lock and implement a pretty aggressive test
that tries to exercise races between reconfigurations and mutations.
Getting the test right took a lot of effort, but the current version
was able to reproduce the problem consistently on macOS; I was unable
to hit the problem on Linux.

PiperOrigin-RevId: 171952377