Handle EOFs properly in reconfiguration's input.
The previous code would enter an infinite loop when the reconfiguration's
input channel reached EOF.  Make sure that this does not happen by handling
EOF appropriately and add a test.

This fixes a problem whereby sandboxfs would enter a tight loop in the
reader goroutine on EOF, causing it to unnecessarily consume a lot of CPU.
(I suspect this might also have caused stuckness problems during exit but
I cannot think of any that I can reproduce.)  Unfortunately, testing for
this is too hard, so the tests ensure a functional sandboxfs even after
EOF and a clean shutdown.

PiperOrigin-RevId: 172084691