Remove the bogus DeleteUnderlyingRoot test.
I can't really tell why the DeleteUnderlyingRoot has been working until
now, but with upcoming changes, the test breaks.  Conceptually, this
test is broken because the kernel is free to cache the directory contents
after a read... and if that happens, there is no way for the kernel to
know that the underlying root directory is gone until the cached contents
expire.  Therefore, remove the test.

PiperOrigin-RevId: 172256353