Rely on in-kernel permission checks for file accesses.
Enable the "default_permissions" FUSE flag to delegate access checks to
the kernel.  With this flag, the kernel performs access decisions based
on the file's ownership and mode without further calling back into
user-space.  This is sufficient for the semantics of sandboxfs.

This yields a significant performance win on macOS because OSXFUSE likes
to flood the user-space file system with Access operations otherwise,
and serving those is costly.  Linux doesn't seem to suffer from this
problem though, and I haven't deeply investigated why.

While doing this, remove the broken custom Access implementation: the
operation was implemented on the BaseNode, not on the invididual node
types... which means that the Go FUSE library never saw this function
and just replied with a default response to the kernel.

And to ensure that all this works as intended, add a bunch of tests for
access permissions.

PiperOrigin-RevId: 172513928