Improve naming of file system node types.
When the dirCommon interface was introduced to represent all possible
directory node types within sandboxfs, this name was chosen to avoid
having to rename Dir to something else.

To avoid confusion (which is a very real thing now that I have to
improve the way Root works), the time has come to fix this naming
issue.

This change starts by renaming Dir to MappedDir (which is the specific
type that represents a directory backed by a physical entity), and then
renaming dirCommon to Dir, which is the interface and thus deserves to
own the generic name.

As part of this change, also rename the generic File and Symlink to
MappedFile and MappedSymlink for consistency with the new MappedDir.
And rename VirtualDir to ScaffoldDir, which better illustrates the
point of these "virtual directories": the previous VirtualDir name was
confusing because everything in a FUSE file system is virtual and thus
the name can lead to confusion to newcomers.

PiperOrigin-RevId: 172778559