Attempt to fix races during reconfigurations.
Add locking to the Root node to handle reads and updates to its backing
directory in a controlled manner.  This fixes obvious problems in e.g.
Rename(), but also potentially fixes more subtle issues that can arise
by the fact that we cannot assume that pointers are atomic (as that's
not part of the Go language specification).

Unfortunately, I do not have a deterministic and simple test case to
expose a problem with the previous code.  What I can say, however, is
that this change allows Bazel to self-build itself with sandboxfs
backing action sandboxing.  I haven't yet been able to expose any kind
of stuckness, but the previous code would show erratic (and random)
behavior.

PiperOrigin-RevId: 172877429