Fix ReadDir operations when the directory is kept open.
If a directory is kept open (by any process!), the kernel may decide to
reuse that open handle to issue the ReadDir requests.  That's fine...
except for the fact that repeated ReadDir operations on a handle do not
start from the beginning of the directory: they start from the current
read position.  As a result, we were returning incomplete directory reads
in this case.

To fix this, simply rewind the directory handle before reading from it.
But only do so on non-first reads to avoid an unnecessary Seek system
call for the vast majority of directory reads.

I discovered this by pure chance: while trying to reproduce another
problem on the command line, I noticed that two consecutive "ls" commands
on the same mount point caused the second one to return no data.  At
first I though this was an issue specific to the Root directory node...
but upon further inspection this proved to be wrong.  The problem was
that macOS's fseventsd daemon opens a file on every mounted file system
on the root directory, and that open file handle is what triggered this
problem.  To make this seem more random, I only hit this because I was
mounting the test instance with --allow=root; otherwise, fseventsd
cannot create its file and the problem doesn't surface.

PiperOrigin-RevId: 172943372