Fix file removals and add tests.
Upon code inspection, I found that file/directory removals were bogus:
a remove operation would always try to remove an underlying file even if
that file had been clobbered by a mapping or by a scaffold directory.

Fix this by explicitly checking for the existence of the entry to be
removed in the various maps we maintain and handing each case
appropriately.

PiperOrigin-RevId: 173151375