Propagate full FileInfo details instead of just node IDs.
An upcoming change will start keeping the full node information details
for underlying files in memory so that the set of f* system calls can be
implemented.  Prepare for such change by propagating those details all
the way to the node creation so that they are easily available when
needed.

PiperOrigin-RevId: 173157761