Add support for operations on file handles with no links left.
Properly implement support to handle operations that update a node's
metadata through open descriptors even when all links to such nodes
are gone.  In particular, this fixes fchmod, fchown, fstat, and
futimes on deleted files.

Discovered the lack of this functionality because macOS's ar(1) tool
fails to work within sandboxfs when TMPDIR is pointed into the sandbox.
But I expect all kinds of things to break without this feature.

PiperOrigin-RevId: 173167220