Disable the "local volume" mount option.
This option is macOS-specific and tells the Finder and other system
services that sandboxfs is a local volume.  When doing so, macOS
has the habit of trying to do things on the volume, like index all
of its contents, and also tries to access hidden files to maintain
extra metadata.  All this additional traffic is costly and it's most
likely of no interest for sandboxfs, so disable it.

On my 3.6GHz 6-core Mac Pro, this cuts down Bazel's self-build time
with sandboxfs enabled from 245 seconds to 215 seconds.

PiperOrigin-RevId: 173167679