Get rid of underlying IDs.
Underlying IDs are a device/inode number pair that uniquely represent
the files backing sandboxfs nodes.  These are used to check, at directory
lookup time, if an entry has changed since the last time it was queried.

The problem is that this is inherently racy.  A layered FUSE file system,
like sandboxfs is, cannot be made aware of changes to the underlying
file system in an atomic manner (because the kernel does not offer such
an interface).  As a result, these underlying ID checks were just reducing
the window between when changes to the underlying file system were made
and when they showed up in sandboxfs -- but they did not eliminate the
race.

Note that there are many more races than this, and note that FUSE already
provides a mechanism to cope with them: node validity times, which we set
to one minute.  We'll be moving towards relying on these instead of
trying to eagerly check for changes, as this will allow more optimizations.

Therefore, remove the tracking of underlying IDs.  This paves the way for
removing unnecessary (and costly) stat operations that sandboxfs currently
issues.  In particular, because this removes the need for a device/inode
pair at node creation time (information that is only available through a
stat system call), this will permit converting the directory read path to
issue only readdir system calls instead of full stats per entry.

PiperOrigin-RevId: 173271473