Preserve node identities across reconfigurations.
Add a cache to track the correspondence of sandboxfs in-memory nodes to the
files that back them.  This is useful to preserve the identity of a node
(i.e. its inode number) across reconfigurations, which also avoids the need
to flush precious kernel node and content caches during reconfigurations.

On my 3.6GHz 6-core Mac Pro, this cuts down Bazel's self-build time
with sandboxfs enabled from 215 seconds to about 200 seconds.

PiperOrigin-RevId: 174624577