Add a Bazel build
This change adds support for Bazel as the preferred build system.  This
is for two reasons: first, to explicitly track the versions of the
dependencies we use, which leads to reproducible builds; and second,
to pave the way for the addition of further development tools.

Note that this does not remove support for the standard Go tools, but
these become not recommended.