Add VSCode configuration
This extends the configure script to support an --enable-vscode flag that
takes care of generating a workspace settings.json file for Visual Studio
Code's benefit.

This generated file points to the correct location of the Go tools as
determined by Bazel and also points to a fake GOPATH generated using the
bazel-gopath tool.