Add source code linting
This adds a script to verify that the code complies with the gofmt and
golint recommendations.

In addition, this adds a Git pre-commit hook to ensure that any new
contributions match this expected style.  Unfortunately, there is no way
to ship this script pre-configured so that "git clone" does the right
thing.  For this reason, add a "configure" script that prepares the
source tree for further contributions.