Implement a --version flag (#10)
Add a version number to sandboxfs and implement a --version flag to dump
it to stdout.  This will be used by Bazel to determine which protocol to
use to communicate with the sandboxfs daemon, allowing us to change the
command-line interface in the future.

Personally, I don't like --version being a flag: given that we have a
subcommand-based interface, this should really be a "version" command.
However, we already have --help so --version doesn't hurt much more.
And I have ideas to kill the subcommand-based interface, which would
homogenize things.