Fix readdirs on root during reconfigurations
The root directory is special because it can change identity during a
reconfiguration.  We solve this by wrapping the backing directory node
type.

Unfortunately, we were not handling open directory handles accordingly:
if the identity changed, an open handle would remain pointing to the
old instance of the backing directory for root.  Combined with the fact
that the kernel seems to reuse open directory handles across operations,
we could get bogus results.

The way this manifested itself is on macOS: the Finder seems to keep an
open handle on the root directory of the file system, and such handle
would cause reconfigurations to never expose new directory listings.
Lookups worked fine, however, because those are handled at the node
level, not the handle level.