Handle transient unmount errors at test teardown (#11)
It is possible, although rare, for a system component such as macOS's
Finder to interfere with the sandboxfs test instances we set up.  One
of the interferences we observe is that the Finder may access the
file system while we are running the test teardown routine, and thus
the unmount we issue fails due to a "resource busy" error.

Under those conditions, the only thing we can do is retry the unmount
operation because the fact that it failed is not our fault.

This should fix some very rare test flakiness I've observed on Travis
and also resolves some consistent flakiness I was observing in new
tests I'm adding, which left me puzzled for a long time.