Combine --{read_only,read_write}_mapping into --mapping
Upcoming changes to modify the directory representation will require
ordering to be preserved across mappings, which is not easy (due to the
flags API in go) if we have multiple flags to represent them.

To resolve this, just merge the mapping representation into a single
flag that also carries the type within its value.