Unify the representation of directories into a single node type
The previous split of MappedDir vs. ScaffoldDir was very complex and got
in the way of upcoming changes to make reconfigurations less intrusive.

Merging the implementation of ScaffoldDir into MappedDir shows that we
can represent a directory in a much simpler way: a way that has a single
mapping of directory entries to nodes, not a layered implementation with
three different maps.

Granted: the new representation is less flexible than the previous one
and requires mappings to be provided in order by the user, but this
restriction feels natural anyway and the gained simplicity is worth it.

Note that, while ScaffoldDir is gone, I'm leaving MappedDir's name
unmodified to keep the churn low in this change.  I'll get rid of the
"mapped" prefix later on.