Improve runfiles handling for lint and install
Change the lint and install tools to use the APIs in the Go rules package
to find their runfiles trees and built binaries within them.

This extra logic is required to later deal with a change in the Go rules
that causes the paths to built Go binaries to not be deterministic
(https://github.com/bazelbuild/rules_go/issues/1239).