Drop the subcommand-based interface
Merge the static and dynamic commands into one, which allows us to drop the
subcommand-based interface.  There never was a good reason to split the
behavior: we can just say that a sandboxfs instance starts running with the
mappings provided on the command line and that it can later be reconfigured
at will via the input/output files.

I'm intentionally not changing the flag names nor the general behavior of
reconfigurations to keep this change simple: fixing callers of sandboxfs
(e.g. all the integration tests) is as simple as dropping the "static"
and "dynamic" subcommand names from the calls.  Moving onto a different
interface is desirable (gRPC could possibly be a nice fit) but, if done,
will come later.

This is a massive simplification that will allow for upcoming changes to
the way the reconfiguration protocol works.