Split the reconfiguration protocol into individual map/unmap operations
The previous sandboxfs exposed a mechanism to change the whole file
system contents "at once" (though not atomically).  This was problematic
because sandboxfs couldn't make any assurances about items being
replaced.

What's worse is that the previous approach made it very hard for Bazel
to share a single sandboxfs instance for different actions.  We'd like
to have each action run in a subdirectory of sandboxfs, but changing
the views of these directories without interfering others was not
possible.  (Bazel could have maintained a full list of all the mappings
and then fed this whole list to sandboxfs for each action... but that's
costly, and due to our poor reconfiguration algorithm, this caused
interference for entries that were supposed to remain unchanged.)

To make this all simpler, change the reconfiguration protocol so that
the user feeds a series of ordered steps to map and unmap independent
entries.  For Bazel, this trivially allows using different top-level
subdirectories for each action.  For other users... this may not be
great, but we don't care for now.  If we want to offer full
reconfigurations at a later stage, we can do so again.