Respect intermediate directory components in nested mappings
If the user maps /foo/bar/baz into the sandbox and /foo/bar already exists
on disk as a descendent of the mapping for root, we should respect both
/foo's and /foo/bar's writability properties.  Do this by transparenty
turning them into mappings to their underlying directories.

We have to do this because that's what the user may expect.  For example:
Bazel creates a directory tree first and then maps individual files as
read-only within that tree, but it expects the rest of the tree to remain
writable as configured on the root node.

This fixes a regression introduced by the unification of directory node
types in PR #14.