Silence eager retry warnings (#20)
It is normal for sandboxfs to take a few milliseconds to start serving.
Our retry code, however, was too eager in printing a warning when the
file system was not up and serving ~immediately, which resulted in the
all test logs containing useless messages.  Fix this by skipping the
warning for the first few attempts.