Add a trivial root directory implementation
This change adds the foundations of the in-memory representation of the
contents of a sandboxfs instance.  In particular, a Node trait that will
handle all node-related operations and the basics of a Dir
implementation.

The implementation of the root directory does nothing interesting at
this point: it just exposes the . and .. entries and returns ENOENT for
any lookups, which is sufficient to introduce the basic file system
concepts.

Note that this does not yet allow us to enable more tests in
travis-build.sh, so the new code introduced by this change was purely
tested locally by mounting the file system and ensuring an "ls -a" shows
the dot entries and that accessing any file returns ENOENT.