Implement parsing of mappings specified in flags
This change adds support for a repeated --mapping flag.  Each instance
of this flag represents a path entry to map within the sandbox.

Note, however, that this only adds the flag but that the value of the
flag is currently unused.  This is a prerequisite for implementing
support for a mapped root directory, which will come next.