Homogenize application of blacklist to input files
Handle the way input files are blacklisted in the exact same way, no
matter if they are discovered or explicitly provided on the command
line.

This is simpler and fixes some issue I've encountered while running
the linter on the whole tree, where it wouldn't skip Rust's target
build directory.