Implement getattr on directories (#28)
This change starts to add support for directory nodes backed by other
directories in the file system, but only implements getattr on them so
far.  The reason is that the logic to manipulate stat data is complex
enough on its own so submitting it as a separate change will help with
the review process.  As a result, this change is pretty much useless
on its own.

I have tested that this works by mapping an arbitrary directory to the
root of the sandbox and ensuring that "stat" on it yields the desired
data.  We'll be able to enable more comprehensive integration tests
once the implementation of the directory is more complete.