Add a cache for nodes and barebones file/symlink nodes
This change adds a factory for nodes in the form of a cache.  The cache
tracks nodes that map to specific underlying files across different
reconfigurations so that we don't lose precious in-kernel state for
files that may be mapped and unmapped many times.  This is specifically
useful in the context of a Bazel build where an action may need to map
the C++ compiler, the following may not need it, but a next one may
need it again: if we don't maintain the node, the kernel will think the
compiler is a different file each time and will have to reload it from
disk.

To make the cache functional, this change also adds barebone
implementations for all other node types we will need, namely File and
Symlink, and adds a "writable" property to each node.

Note, however, that this code is pretty much useless as it is because
the only call we have now is to create the node for the mount point.
This will become useful in the next commit, where I'll implement
read-only mapped directories (i.e. their readdir and lookup operations).