Make File/Symlink's underlying_path non-optional
Fix the representation of file and symlink nodes to have an underlying
path at all times; it needn't be optional.

The fact that it was previously optional was just an overlook caused
by the rather weird get_new_attr() helper function that I added in an
attempt to "simplify" the implementation of getattr().  This function
is now gone.