Properly propagate and display IO error causes
Change the code in lib.rs to propagate IO errors as failure::Error
so that they can be annotated with context, and then change the
main program to concatenate all causes when displaying the error
message.

This is necessary to implement multiple mappings as we will encounter
a variety of errors at different stages and each must carry its own
context.