Add handles for directories
The previous implementation used a single readdir hook to hande reads
from directories but this prevents supporting the case of reading from
an open-but-deleted directory -- which we need to support.

To fix this, add handle tracking to directories and implement separate
opendir/readdir operations for them.

There are integration tests that cover this functionality, but they are
coupled with the presence of setattr.  That's how I discovered that I
had to implement directory handles, but this means that I cannot enable
those tests until setattr lands (next PR).