Add placeholder implementations of RW operations
This essentially replaces the default implementation of all RW
operations we expect to implement, which returns ENOSYS, for a
version that panics.

The rationale for doing this is to make debugging easier.  When
implementing setattr, I spent a lot of time trying to understand
why some tests I thought should be passing weren't, and that was
because they were invoking an unimplemented API.  If sandboxfs
had panicked, the stacktrace would have immediately pinpointed
the problem.