Fix error codes for writes to unwritable nodes
Nodes marked as non-writable cannot be modified regardless of who the
calling user is... so the correct error to return is EPERM, not EACCES.
The difference is important, but the error names are so subtle that
it's too easy to get them wrong (and I did).

This fixes an error code during setattr and another one when attempting
to open files for write.