Prepare to support setattr on deleted files
In order to implement unlink and rmdir, which will come later, we must
prepare to deal with the case of applying updates to a delete node via
still-open handles.

To do this, avoid doing a getattr immediately after a setattr (because
it may not be possible once the underlying file is gone), and instead
compute what the updated attributes should be.  This is not completely
accurate because of ctime, but it's good enough for now.

Also, make the input path to setattr optional so that we can ask this
function to only update the in-memory properties.