Return EOPNOTSUPP when trying to setattr the times of a symlink
The setattr hook should be using futimensat to update the timestamps
of the node so that we could support the case of modifying those of
a symlink (without following it).

Unfortunately, we currently cannot do so because our testing environment
does not have this system call available in all platforms... so just
return EOPNOTSUPP for now (instead of doing the wrong thing and following
the symlink).