Make SandboxFS movable across threads
The SandboxFS type included multiple references to the Node and Handle
traits without Send and Sync annotations, which prevented it from being
shared across threads.  We need this type to be Send+Sync because we
will need to use at least one extra thread to handle the reconfiguration
loop.

To fix this, introduce two type aliases, ArcNode and ArcHandle, that
define all necessary traits for defining the corresponding node and
handles in a way that is thread-friendly, and then use those everywhere.