Add a file type with one owner but many viewers
In order to implement reconfigurations, we will need to open a file to
read reconfiguration requests from and such reads will have to happen
from a separate thread than the one that blocks on fuse::mount.

Regardless of the status of such thread (which will typically be blocked
on a read), we want to ensure that sandboxfs can be unmounted at any
point even if the reconfiguration input file (possibly a FIFO) hasn't
been closed yet.  This means that the we must be able to tell the
reconfiguration thread to exit as soon as fuse::mount returns.

The simplest way to do this is to close the file handle on which the
reads in the reconfiguration thread are blocked.  Unfortunately there
doesn't seem to be anything in the Rust standard API supporting this.
fs::File handles can be cloned but doing so results in a handle with
multiple owners and no ability to explicitly close the underlying file
descriptor.

To resolve this, add our own simple "ShareableFile" wrapper over a
raw file descriptor that supports a single owner (with close semantics)
and multiple asynchronous readers.